add error messages

+++ Command/clear.js
@@ -6,7 +6,11 @@ module.exports.run = (client, message, args) => {
 
 	if (message.member.hasPermission("ADMINISTRATOR")) {
         let arguments = args.join("")
-        message.channel.bulkDelete(arguments).then(messages => message.channel.send(arguments + " messages supprimés"))
+        if(!arguments){
+            message.channel.send("Tu dois me donner un nombre de messages a supprimer stp !")
+        }else{
+            message.channel.bulkDelete(arguments).then(messages => message.channel.send(arguments + " messages supprimés"))
+        }
     } else {
         message.author.send("Tu n'a pas la permission de faire cet commande !")
     }
